Message Method Redraws And Updates
The message methods includes the for loop that redraws the current state
of the board and the call to the pygame update method.  This is intended
to eliminate the re-use of code that was necessary to display messages
properly.

This commit also includes the compiled cython code that came with the
original code I got.