Message Update And Button Class
Updated the message method to allow setting the coordinates of the
message box relative to the game screen.

Added a button class that uses the same display method as the message
method, but with the background and text colors inverted to
differentiate them.  The button class also allows functions to be passed
as arguments, which are then triggered when the button's run method is
executed.  The is_clicked method only returns true or false, as to allow
flexibility in determining certain conditions before committing to
running the attached function.