Minor Button Update And MultiplayerServer Class
Only 2 buttons are stored in the buttons variable since they will be the
first buttons the player sees when starting the game.  Additional
buttons will be appended to the list as needed during runtime.

The multiplayer_serv.py MyClass has been changed to the
MultiplayerServer class to clarify what is being called when the class
gets called to start the multiplayer server.