Button Class Changes 2
Attempted to create the start menu with single and multi player buttons
that allow the player to choose if they are playing alone or with others
before the game starts.  The buttons display properly, but the function
of the buttons still needs work, as the game crashes upon clicking on a
button.

Made changes to the gameexit method to allow an array of tuples
containing the text and the delay time before the next message in the
list displays.